+++ src/SMAPI.Web/ViewModels/IndexModel.cs
@@ -6,17 +6,11 @@ namespace StardewModdingAPI.Web.ViewModels
         /*********
         ** Accessors
         *********/
-        /// <summary>The latest SMAPI version.</summary>
-        public string LatestVersion { get; set; }
+        /// <summary>The latest stable SMAPI version.</summary>
+        public IndexVersionModel StableVersion { get; set; }
 
-        /// <summary>The Markdown description for the release.</summary>
-        public string Description { get; set; }
-
-        /// <summary>The main download URL.</summary>
-        public string DownloadUrl { get; set; }
-
-        /// <summary>The for-developers download URL.</summary>
-        public string DevDownloadUrl { get; set; }
+        /// <summary>The latest prerelease SMAPI version (if newer than <see cref="StableVersion"/>).</summary>
+        public IndexVersionModel BetaVersion { get; set; }
 
 
         /*********
@@ -26,16 +20,12 @@ namespace StardewModdingAPI.Web.ViewModels
         public IndexModel() { }
 
         /// <summary>Construct an instance.</summary>
-        /// <param name="latestVersion">The latest SMAPI version.</param>
-        /// <param name="description">The Markdown description for the release.</param>
-        /// <param name="downloadUrl">The main download URL.</param>
-        /// <param name="devDownloadUrl">The for-developers download URL.</param>
-        internal IndexModel(string latestVersion, string description, string downloadUrl, string devDownloadUrl)
+        /// <param name="stableVersion">The latest stable SMAPI version.</param>
+        /// <param name="betaVersion">The latest prerelease SMAPI version (if newer than <paramref name="stableVersion"/>).</param>
+        internal IndexModel(IndexVersionModel stableVersion, IndexVersionModel betaVersion)
         {
-            this.LatestVersion = latestVersion;
-            this.Description = description;
-            this.DownloadUrl = downloadUrl;
-            this.DevDownloadUrl = devDownloadUrl;
+            this.StableVersion = stableVersion;
+            this.BetaVersion = betaVersion;
         }
     }
 }